requirements.txt
-i https://pypi.org/simple
asgiref==3.11.0; python_version >= '3.9'
dj-database-url==3.0.1
django==5.2.8; python_version >= '3.10'
django-admin-honeypot-updated-2021==1.2.0; python_version >= '3.6'
django-browser-reload==1.21.0; python_version >= '3.9'
django-cleanup==9.0.0; python_version >= '3.8'
django-debug-toolbar==6.1.0; python_version >= '3.9'
django-environ==0.12.0; python_version >= '3.9' and python_version < '4'
django-extensions==4.1; python_version >= '3.9'
django-ipware==7.0.1; python_version >= '3.8'
django-redis==6.0.0; python_version >= '3.9'
psycopg2-binary==2.9.11; python_version >= '3.9'
python-ipware==3.0.0; python_version >= '3.7'
redis==7.1.0; python_version >= '3.10'
sqlparse==0.5.3; python_version >= '3.8'
tzdata==2025.2; python_version >= '2'
